backend/server/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
exceptiongroup==1.3.1x
ffmpeg-python==0.2.0
filelock==3.20.3
fsspec==2026.1.0
future==1.0.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.4.2
numba==0.63.1
numpy==2.2.6
openai-whisper==20250625
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yt-dlp==2025.12.8
